requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.48
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
pillow==12.1.0
python-multipart==0.0.20
aiofiles==24.1.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
boto3==1.38.24
python-jose[cryptography]==3.3.0
pytest==8.4.1
pytest-asyncio==1.0.0
